requirements.txt
streamlit==1.31.1
attr==0.3.2
ConfigParser==7.0.0
contextlib2==21.6.0
cryptography==42.0.8
Cython==3.0.10
dl==0.1.0
docutils==0.21.2
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.3
Jinja2==3.1.3
jnius==1.1.0
keyring==25.2.1
lockfile==0.12.2
mock==5.1.0
numpy==2.0.0
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
Pillow==10.1.0
Pillow==10.3.0
protobuf==3.20.3
pyOpenSSL==24.1.0
pywin32==306
railroad==0.5.0
Sphinx==7.3.7
thread==2.0.3
tornado==6.4
trove_classifiers==2024.5.22
truststore==0.9.1
ultralytics==8.2.12
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
torchvision==0.17.0
opencv-python-headless
